@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "start": "ng serve --delete-output-path=false",
+    "start": "ng serve",
     "build": "node build.js",
     "test": "karma start",
     "install-package": "npm run build && npm pack ./dist && npm install ngx-ui-scroll-0.0.4.tgz --no-save",
@@ -11,24 +11,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.7",
-    "@angular/common": "^5.2.7",
-    "@angular/core": "^5.2.7",
-    "@angular/forms": "^5.2.7",
-    "@angular/http": "^5.2.7",
-    "@angular/platform-browser": "^5.2.7",
-    "@angular/platform-browser-dynamic": "^5.2.7",
-    "@angular/router": "^5.2.7",
+    "@angular/animations": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/router": "6.0.2",
     "core-js": "^2.5.3",
-    "ngx-bootstrap": "^2.0.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.20"
+    "ngx-bootstrap": "^3.0.0",
+    "rxjs": "^6.1.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler": "^5.2.7",
-    "@angular/compiler-cli": "^5.2.7",
-    "@angular/language-service": "^5.2.7",
+    "@angular/cli": "6.0.3",
+    "@angular/compiler": "6.0.2",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular/language-service": "6.0.2",
+    "@angular-devkit/build-angular": "~0.6.3",
     "@types/jasmine": "^2.5.54",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.95",
