@@ -1,4 +1,4 @@
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Settings, DevSettings } from './settings';
 import { Adapter } from './adapter';
 
