@@ -1,5 +1,4 @@
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
+import { Observable, Observer } from 'rxjs';
 
 import { Scroller } from '../scroller';
 import { Process, ProcessSubject } from '../interfaces/index';
